Improve test configuration
A container is only started in an integration test scenario so the
"server.port" property is actually useless in most modules.

There is no reason to force the use of MySQL in tests as an embedded
database can perfectly do the job. H2 has been chosen but HSQL and Apache
Derby are supported as well.

BasicFlowTest is actually a test that requires the container. Rather than
starting the app ourselves, we create an integration test with a random
port and inject the actual allocated port in our test. That way, the test
can run without forciing all users to free a given port.