Clean maven config
Dependency versions should not be copy/pasted around. Instead, the "main"
pom of the project should have a `dependencyManagement` section that
defines the dependencies the project uses.

If jetty should be used instead of tomcat, it should be excluded rather
than being flagged as provided.

Runtime dependencies should be marked accordingly as well. Note that if
the test improvements are applied first, the mysql and hikari definitions
should move to the main application. And it makes a lot more sense that
way.